+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Added
+
+- Add current context's namespace as `:current_namespace` field to Kubeconfig.
+- Add Req option `:context` to switch context ad-hoc.
+
 <!--------------------- Don't add new entries after this line --------------------->
 
 ## 0.3.0 - 2024-10-04